Refactor chat item displays
